Refactoring elements in the AuctionHouseClient file
